Make it so that userEvents contains only Event.js files
