+++ lib/github/core/EventTemplates.js
