+++ lib/github/core/HtmlWrapper.js
@@ -1,7 +1,7 @@
 const renderStringWithData = require('./renderHelpers').renderStringWithData;
 const EventTemplates = require('./EventTemplates');
 const EventTypes = require('./EventTypes');
-const { hrefForRepo, hrefForBranch, link } = require('./eventLinks/core');
+const { hrefForRepo, hrefForBranch, link } = require('./linkUtils');
 
 class HtmlWrapper {
