+++ lib/github/core/describeTime.js
