+++ lib/github/core/fetchUserEvents.js
