+++ lib/github/core/linkUtils.js
