+++ lib/github/core/renderHelpers.js
