+++ lib/github/userEvents/BaseEvent.js
@@ -1,4 +1,4 @@
-const describeTime = require('../describeTime');
+const describeTime = require('../core/describeTime');
 
 class BaseEvent {
