+++ lib/github/userEvents/CreateEvent.js
@@ -1,5 +1,5 @@
 const BaseEvent = require('./BaseEvent');
-const { hrefForRepo, hrefForBranch } = require('./eventLinks/core');
+const { hrefForRepo, hrefForBranch } = require('../core/linkUtils');
 
 class CreateEvent extends BaseEvent {
