+++ lib/github/userEvents/PushEvent.js
@@ -2,7 +2,7 @@ const BaseEvent = require('./BaseEvent');
 
 const removeRefsHeads = (ref) => ref.substring('refs/heads/'.length);
 
-const { hrefForRepo } = require('./eventLinks/core');
+const { hrefForRepo } = require('../core/linkUtils');
 
 class PushEvent extends BaseEvent {
